Various and sundry
