+++ params.toml
@@ -1,11 +1,24 @@
 # 02-712 Project Pipeline Settings
 
+[data]
+subsampling = 0.05
+
 [training]
 latent_dimensions_initial_guess = 32
 newton_temperature = 1
 batch_size = 256
 
-[training.vae_trainer]
-# See: https://pytorch-lightning.readthedocs.io/en/latest/trainer.html
-max_epochs = 2
-gpus = 0
+    [training.vae_trainer]
+    # See: https://pytorch-lightning.readthedocs.io/en/latest/trainer.html
+    max_epochs = 2
+
+[simulation]
+minimum_cells_per_class = 2
+
+[classifier]
+n_hidden_layers = 2
+
+    [classifier.training_opts]
+    # See: https://pytorch-lightning.readthedocs.io/en/latest/trainer.html
+    fast_dev_run = false
+    max_epochs = 10
